Punishment sided
Reworked ;warns to actually show proper warnings from both sides
Created ;suspendRemove to remove a suspension from the files (No active suspensions)
Created ;punishments to show all punishments inside of 1 command
Reworked ;warnremove to not use indexing, but rather use button indexing to be easier to use.

Edited 3 things inside of extensions.js
- message.confirmButton: Will no longer set the buttons to inactive since that will sometimes cause a error
- message.confirmNumbers: Now supports up to 24 options instead of 10
- guild.findMember: Now can find a member easier, mentions, id or nickname.

Made ;log actually delete the confirm message

+++ commands/warnRemove.js
@@ -1,38 +1,30 @@
 const Discord = require('discord.js')
 const ErrorLogger = require('../lib/logError')
+const moment = require('moment');
 
 module.exports = {
     name: 'warnremove',
     alias: ['removewarn'],
     description: 'Removes warn from user',
-    args: '<user> <warn number>',
+    args: '<user>',
     requiredArgs: 1,
     role: 'security',
     async execute(message, args, bot, db) {
-        if (args.length < 2) message.channel.send('Command Entered incorrectly. Please try again')
-        let member = message.mentions.members.first()
-        if (!member) member = message.guild.members.cache.get(args[0])
-        if (!member) member = message.guild.members.cache.filter(user => user.nickname != null).find(nick => nick.nickname.replace(/[^a-z|]/gi, '').toLowerCase().split('|').includes(args[0].toLowerCase()));
+        let member = message.guild.findMember(args[0])
         if (!member) return message.channel.send('Member not found. Please try again')
-        let toRemove = parseInt(args[1])
-        db.query(`SELECT * FROM warns WHERE id = '${member.user.id}'`, async function (err, rows) {
+        db.query(`SELECT * FROM warns WHERE id = '${member.user.id}' AND guildid = '${message.guild.id}'`, async function (err, rows) {
             if (err) ErrorLogger.log(err, bot)
-            let warn = rows[toRemove - 1]
-            if (!warn) return message.channel.send(`Warn number ${toRemove} was not found. Please try again`)
-            if (warn.guildid != message.guild.id) return message.channel.send('That warning was not issued in this server. Please retry')
-            let confirmEmbed = new Discord.EmbedBuilder()
-                .setColor('#ff0000')
-                .setTitle('Please Confirm')
-                .setDescription(`__Warn for user:__ ${member} (${member.nickname})\n__Reason:__${warn.reason}\n__Warn by:__ <@!${warn.modid}>`)
-            await message.channel.send({ embeds: [confirmEmbed] }).then(async confirmMessage => {
-                if (await confirmMessage.confirmButton(message.author.id)) {
-                    db.query(`DELETE FROM warns WHERE warn_id = ${warn.warn_id} AND modid = '${warn.modid}'`)
-                    confirmEmbed.setTitle('Warn Removed')
-                    confirmMessage.edit({ embeds: [confirmEmbed], components: [] })
-                } else {
-                    confirmMessage.delete()
-                }
-            })
+            for (let i in rows) { let index = parseInt(i); rows[i].index = index}
+            let embed = new Discord.EmbedBuilder()
+                .setColor('#F04747')
+                .setTitle('Confirm Action')
+                .setDescription(rows.map(warning => `${warning.index+1}. By <@!${warning.modid}> ${moment().to(new Date(parseInt(warning.time)))}\`\`\`${warning.reason}\`\`\``).join('\n'))
+            let confirmMessage = await message.channel.send({ embeds: [embed] })
+            const choice = await confirmMessage.confirmNumber(rows.length, message.member.id);
+            if (!choice || isNaN(choice) || choice == 'Cancelled') return await confirmMessage.delete();
+            let removeWarning = rows[choice - 1]
+            db.query(`DELETE FROM warns WHERE warn_id = ${removeWarning.warn_id} AND modid = '${removeWarning.modid}'`)
+            await confirmMessage.delete()
         })
     }
 }
